+++ CovidClassModel.ucls
@@ -1,17 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <class-diagram version="1.2.4" icons="true" always-add-relationships="false" generalizations="true" realizations="true" 
   associations="true" dependencies="false" nesting-relationships="true" router="FAN">  
-  <enumeration id="1" language="java" name="br.edu.ufcg.covidlog.user.Role" project="covid-log-back" 
-    file="/covid-log-back/src/main/java/br/edu/ufcg/covidlog/user/User.java" binary="false" corner="BOTTOM_RIGHT">    
-    <position height="-1" width="-1" x="121" y="112"/>    
-    <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
-      sort-features="false" accessors="true" visibility="true">      
-      <attributes public="true" package="true" protected="true" private="true" static="true"/>      
-      <operations public="true" package="true" protected="true" private="true" static="true"/>    
-    </display>  
-  </enumeration>  
-  <class id="2" language="java" name="br.edu.ufcg.covidlog.user.User" project="covid-log-back" 
-    file="/covid-log-back/src/main/java/br/edu/ufcg/covidlog/user/User.java" binary="false" corner="BOTTOM_RIGHT">    
+  <class id="1" language="java" name="br.edu.ufcg.covidlog.model.User" project="covid-log-back" 
+    file="/covid-log-back/src/main/java/br/edu/ufcg/covidlog/model/User.java" binary="false" corner="BOTTOM_RIGHT">    
     <position height="-1" width="-1" x="332" y="114"/>    
     <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
       sort-features="false" accessors="true" visibility="true">      
@@ -19,33 +10,24 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <class id="3" language="java" name="br.edu.ufcg.covidlog.entity.Entity" project="covid-log-back" 
-    file="/covid-log-back/src/main/java/br/edu/ufcg/covidlog/entity/Entity.java" binary="false" corner="BOTTOM_RIGHT">    
-    <position height="-1" width="-1" x="663" y="124"/>    
+  <class id="2" language="java" name="br.edu.ufcg.covidlog.model.Entity" project="covid-log-back" 
+    file="/covid-log-back/src/main/java/br/edu/ufcg/covidlog/model/Entity.java" binary="false" corner="BOTTOM_RIGHT">    
+    <position height="-1" width="-1" x="585" y="122"/>    
     <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
       sort-features="false" accessors="true" visibility="true">      
       <attributes public="true" package="true" protected="true" private="true" static="true"/>      
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <class id="4" language="java" name="br.edu.ufcg.covidlog.donation.Donation" project="covid-log-back" 
-    file="/covid-log-back/src/main/java/br/edu/ufcg/covidlog/donation/Donation.java" binary="false" 
-    corner="BOTTOM_RIGHT">    
-    <position height="-1" width="-1" x="378" y="308"/>    
+  <class id="3" language="java" name="br.edu.ufcg.covidlog.model.Donation" project="covid-log-back" 
+    file="/covid-log-back/src/main/java/br/edu/ufcg/covidlog/model/Donation.java" binary="false" corner="BOTTOM_RIGHT">    
+    <position height="-1" width="-1" x="383" y="291"/>    
     <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
       sort-features="false" accessors="true" visibility="true">      
       <attributes public="true" package="true" protected="true" private="true" static="true"/>      
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <association id="5">    
-    <end type="SOURCE" refId="2" navigable="false">      
-      <attribute id="6" name="role"/>      
-      <multiplicity id="7" minimum="0" maximum="1"/>    
-    </end>    
-    <end type="TARGET" refId="1" navigable="true"/>    
-    <display labels="true" multiplicity="true"/>  
-  </association>  
   <classifier-display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
     sort-features="false" accessors="true" visibility="true">    
     <attributes public="true" package="true" protected="true" private="true" static="true"/>
