--- src/main/java/br/edu/ufcg/covidlog/entity/Entity.java
@@ -1,14 +1,0 @@
-package br.edu.ufcg.covidlog.entity;
-
-/**
- * Entity is a community or group that demands any help.
- * 
- * @author adalbertocajueiro
- *
- */
-public class Entity {
-	private String name;
-	private String id;
-	
-	//precisa representar aqui as coordenadas do local
-}
