+++ src/main/java/br/edu/ufcg/covidlog/model/User.java
@@ -1,0 +1,8 @@
+package br.edu.ufcg.covidlog.model;
+
+import br.edu.ufcg.covidlog.enums.Role;
+
+public class User {
+	private Role role;
+	private String email;
+}
