--- src/main/java/br/edu/ufcg/covidlog/user/User.java
@@ -1,7 +1,0 @@
-package br.edu.ufcg.covidlog.user;
-
-enum Role {ADMIN,DONOR}
-public class User {
-	private Role role;
-	private String email;
-}
